requirements.txt
Flask==0.10.1

Flask-Migrate==2.0.3

Flask-Sockets==0.2.0

Flask-SQLAlchemy==1.0

Flask-Heroku==0.1.9

gevent==1.0.2

gevent-websocket==0.9.5

greenlet==0.4.9

gunicorn==19.4.5

itsdangerous==0.24

Jinja2==2.8

MarkupSafe==0.23

peewee==2.8.7

psycopg2==2.5.3

redis==2.10.5

Werkzeug==0.11.3
